Refactor Amount: Split between Add and Sub operation
If we want to use Amount consistently accross Sebak, we first need to remove all uses of negative value.
Providing a safe interface to do addition and substraction doing underflow and overflow check is the first step.
In addition, `BaseFee` was typed, and some conversion were removed.

+++ lib/operation_payment.go
@@ -66,7 +66,7 @@ func FinishOperationPayment(st *sebakstorage.LevelDBBackend, tx Transaction, op 
 	}
 
 	var expectedTarget Amount
-	if expectedTarget, err = baTarget.GetBalanceAmount().Add(int64(op.B.GetAmount())); err != nil {
+	if expectedTarget, err = baTarget.GetBalanceAmount().Add(op.B.GetAmount()); err != nil {
 		return
 	}
